Distinguish between TCP connection being "open" and AMQP connection being "open"
According to AMQP 0.9.1 spec (section 2.3.7):

  A connection or channel is considered  open  for the client when it has sent Open,
  and for the server when it has sent Open-Ok. From this point onwards a peer that wishes
  to close the channel or connection MUST do so using a hand-shake protocol
  that is described here.

We definitely did not follow this rule as spec example with incorrect credentials
has revealed (@connected was set to true as early as EventMachine has called us back about
TCP connection setup, even before we've negotiated AMQP protocol version with the broker).

However, from AMQP client user perspective connection is considered open when
we are also authorized to access vhost. That is why I only set @connected to be true
after we receive connection.open-ok from the broker.

And then there is authorization issues handling (for example, user has no permission
to declare a particular exchange or bind a particular queue), which is considered to
be "a connection issue" by some people, but is neither covered by this change, nor
by spec examples we have so far. JFYI.