Let default exchange channel to be passed in; don't cache default exchange object (read on!)
When default exchange object is created with a new channel calling code can't grab reference to,
it is surprising at least. When it turns out that even if the channel has been closed already
(due to connection loss or as part of test suite teardown actions, for example), any action on
that exchange will result in an exception that is very difficult to reason about.

It took me good 20 minutes to investigate why specs were failing for no reason. Performance
"gain" here is negligible compared to amount of confusion this behavior will create on to
library users.

It's Just Not Worth It™.