Add tests related to queue exclusivity (read on)
According to AMQP 0.9.1 spec,

  "Exclusive queues may only be accessed by the current connection,
   and are deleted when that connection closes. Passive declaration
   of an exclusive queue by other connections are not allowed."

So far we only cover #1 clause out of 3, because it seems to be the most
important aspect.

Because attempts to access exclusive queue from another connection
results in a connection-level exception, the best we can do is to verify
that connection is closed by the peer. If you have better ideas about
this specific test case, let us know at http://bit.ly/hjfJrY