Add support for Heartbeat frames.
  - Communicate the heartbeat interval upon Connection::Tune
  - Update @last_server_heartbeat when Heartbeat frames are received
  - Initialize @last_server_heartbeat, otherwise first comparison in heartbeat
    timer will raise an exception (nil < ..)

This is a lot like Uwe Kubosch's 0838b48e, except Time.now is relatively
expensive, AMQP servers will incur that cost heavily when throughput is high,
and I've dealt with one Broker implementation that disconnected me if I failed
to send a Heartbeat frame, even when there was tons of other traffic.

So this patch goes easy on Time.now and only records the time of actual
Heartbeat frames.