Long overdue documentation update for Channel, Exchange and Queue
Apart from making statements that are simply not true (like that messages
delivered to durable exchanges are persisted on disk), old documentation
structure resulted in a lot of duplication between documentation for
methods like Channel#direct and Exchange class. Not only that
duplication is a pain to maintain, it also presented a dilemma for
documentation write:

 => So, where do I put introduction to exchange types and so on?

And so documentation for individual methods tried to explain half
of AMQP spec. Not a very good idea.

What I did removes most of duplication (all of it that matters),
moves introduction & overview to Exchange class documentation and
uses YARD links to hints at introduction part from docs for methods
like Channel#direct.

We also use @see tags extensively on class level and provide
Apple documentation-like sections that describe key methods on
each class.

Note that @example tag seem to only work on methods, so I had to
leave all examples there.