Land connection interruption handler example
Once you run it, stop AMQP broker for a few seconds then bring it back
up and check (broker's) log to see that client actually reconnected.

Note that params order for #reconnect is the opposite of that in amq-client
because we have to keep BasicClient#reconnect backwards-compatible.
amq-client method signature has to change instead.