AMQP::BasicClient => AMQP::Session
Session is not a very good name either but "client" makes even less sense.

In AMQ::Client, connection is handled by two classes: EventMachineClient
and Connection. First former is a TCP connection while the latter is AMQP
connection. AMQP connection may be teared down and re-estabilished without
closing underlying TCP connection. Ok, this is not what we currently support
but you get the idea.

So Session it is, for now. BasicClient in 0.7.x is not a class (and is
used by EventMachine under the hood), so keeping it around is unlikely
to be useful to anyone.