AMQP URIs should not imply a leading slash in the vhost
The amqp:// URI convention currently implemented by the amqp gem has a
problem: It uses the path component of the URI as the vhost name,
including the leading slash.

There is no definitive standard for amqp URIs.  But every other
proposed amqp URI scheme (including the two qpid schemes, and the
one implemented by rabbitmq-shovel). uses the path *without* the
leading slash as the vhost (though they differ in exactly how
they convert the rest of the path to a vhost name).

There is a clear advantage to the more widespread scheme: It can
refer to vhosts that do not begin with a slash (despite the
RabbitMQ default vhost being called '/', there is nothing to
suggest that other vhost names should begin in a slash).  It is
obviously desirable that a URI can be constructured to refer to
any legal vhost.

It's unlikely that the URI connection string is being widely used
so far, so it is a good idea to fix this now.

With this change:

'amqp://host' uses the default vhost '/'

'amqp://host/' uses the vhost with the empty name ''

'amqp://host/foo' uses the vhost named 'foo'

'amqp://host/%2Ffoo' uses the vhost named '/foo'