Root amq-client in the amqp gem repo
This is our first step to getting rid of amq-client.
It was a good idea 2 years ago but now it's clear that
there's next to no value in it for clients that do not
try to mimic amqp gem's API, such as Bunny 0.9.

Bunny and amqp gem already reuse amq-protocol, but their
networking and public API parts are too different to try
to extract a common library.

So long amq-client. You've served amqp gem well for good
2 years.